[REFACTOR] `AssemblePost`를 `Transform` 서비스로 이관
- 게시글을 조립하는 AssemblePost 또한 Transformation이 주 관심사로, 하위 서비스에 있는 것이 바람직
- 테스트 용이성을 위해 하위 서비스로 분할하였음
