+++ backend/src/api/transformation/transformation.service.ts
@@ -9,9 +9,17 @@ import { DecomposedFileDto } from './dtos/decomposed-file.dto';
 import { DecomposedPostDto } from './dtos/decomposed-post.dto';
 import { WritePostDto } from '../post-api/dtos/post/write-post.dto';
 import { ModifyPostDto } from '../post-api/dtos/post/modify-post.dto';
+import { UserPayload } from '@/common/guards/user-payload.interface';
+import { BlockDto } from '@/domain/block/dtos/block.dto';
+import { PostDto } from '@/domain/post/dtos/post.dto';
+import { Post, Block, File } from '@prisma/client';
+import { FileDto } from '../post-api/dtos/file.dto';
+import { UtilityService } from '@/common/utility/utility.service';
 
 @Injectable()
 export class TransformationService {
+  constructor(private readonly utils: UtilityService) {}
+
   private parseLatLon(latLon: string) {
     const [lat, lon] = latLon.split(',').map((s) => parseFloat(s.trim()));
     return { latitude: lat, longitude: lon };
@@ -90,5 +98,53 @@ export class TransformationService {
     });
 
     return plainToInstance(DecomposedPostDataDto, { post: postDto, blocks: blockDtos, files: fileDtos });
+  }
+
+  /**
+   * 게시글, 사용자, 블록, 파일 정보를 받아 계층적인 게시글 데이터 구조로 변환한다.
+   * @param post 게시글 엔티티
+   * @param user 사용자 정보 (닉네임, 이메일)
+   * @param blocks 블록 엔티티 배열
+   * @param files 파일 엔티티 배열
+   * @returns 게시글, 블록, 파일 순으로 계층화된 데이터
+   */
+  assemblePost(post: Post, user: UserPayload, blocks: Block[], files: File[]): PostDto {
+    const blockDtoMap = this.createBlockDtoMap(files, blocks);
+    return PostDto.of(post, user, blockDtoMap.get(post.uuid)!);
+  }
+
+  /**
+   * 여러 게시글의 사용자, 블록, 파일 정보를 받아 계층적인 게시글 데이터 구조로 변환한다.
+   * @param post 게시글 엔티티
+   * @param user 사용자 정보 (닉네임, 이메일)
+   * @param blocks 블록 엔티티 배열
+   * @param files 파일 엔티티 배열
+   * @returns 게시글, 블록, 파일 순으로 계층화된 데이터 배열
+   */
+  assemblePosts(posts: Post[], users: UserPayload[], blocks: Block[], files: File[]): PostDto[] {
+    const blockDtoMap = this.createBlockDtoMap(files, blocks);
+    return posts.map((post, index) => PostDto.of(post, users[index], blockDtoMap.get(post.uuid)!));
+  }
+
+  /**
+   * 블록 및 파일 정보를 받아, Map 자료 구조로 반환한다.
+   * @param files 파일 엔티티 배열
+   * @param blocks 블록 엔티티 배열
+   * @returns 블록과 파일 DTO로 구조와된 Map
+   */
+  private createBlockDtoMap(files: File[], blocks: Block[]) {
+    const fileDtoMap = this.utils.toTransMapFromArray<File, string, FileDto>(
+      files,
+      (file: File) => file.sourceUuid!,
+      (file: File) => FileDto.of(file),
+    );
+
+    const blockDtoMap = this.utils.toTransMapFromArray<Block, string, BlockDto>(
+      blocks,
+      (block: Block) => block.postUuid,
+      (block: Block) => BlockDto.of(block, fileDtoMap.get(block.uuid)),
+    );
+
+    return blockDtoMap;
   }
 }
